refactor slot initialization
Currently, there is a lot of code duplication between the slot
initialization functions for the pool and the slab. There are a number
of similar methods that differ only in what `Slot` method they call,
which is a bit confusing when reading the code, and requires duplicating
a lot of common functionality.

I've refactored this code path so that the same `Shard`, page `Shared`
data, and `Slot` methods are used by both the `Pool` and the `Slab`,
with all differences in behavior (namely, which `Slot` method is called)
specified at the callsite. The new implementation has significantly less
code duplication, and is hopefully much clearer.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>