docs(Slab::take): document that `take` spins
This commit updates the API docs for `Slab::take` to note explicitly
that the calling thread is blocked by spinning until outstanding
references to the removed slot are dropped. The docs now warn against
using `take` if references to a slot may last for a long time, and note
that it should only be used when such references are relatively
short-lived.

Fixes #20

Signed-off-by: Eliza Weisman <eliza@buoyant.io>