feat(Pool): add a sharded object pool for reusing heap allocations (#19)
Add a new API for reusing heap allocation backed by Shards.

This PR adds new APIs to reusing heap allocations. As covered in #2 ,
when using the `Slab`, any heap allocated object stored in the slab
would be dropped when the object was removed from it. This meant that
each object stored in the `Slab` would have to be individually
allocated.  The `Pool` type aims to reuse these allocations by providing
an in-place modification APIs.

In order to provide this, the internals of the crate have been
significantly refactored in order to allow as much code reuse between
the two types as possible. `Pool` specific APIs have the additional
trait bounds of `T: Clear + Default` and methods specific for this type
are in impl blocks bounded by these traits. Slab specific methods are
implemented for types generic over `Option<T>`.  This is a pattern which
is followed down the abstractions to `Slot`. 

Closes #2
Closes #15