fix(Slab): invalid generation in case of custom config (#80)
Hi, I used the following config and found the incorrect calculations of the generation and reference counter. After debugging, I  discovered that sometimes `Generation::pack` is used instead of `LifecycleGen::pack`, which can be fixed by this PR.

My config:
```rust
struct SlabConfig;
impl sharded_slab::Config for SlabConfig {
    const INITIAL_PAGE_SIZE: usize = 32;
    const MAX_PAGES: usize = 15;
    const MAX_THREADS: usize = 256;
    const RESERVED_BITS: usize = 24;
}
```

The bug cannot be reproduced with the default config, because `Generation` and `LifecycleGen` have the same `Pack::SHIFT`.

It would be nice to remove `impl Pack for Generation` in favor of different `impl Pack for KeyGen` and `impl Pack for LifecycleGen`, but I decided not to do any extra refactoring, which you may not accept.

If the patch is ok for you, I can add a test to prevent regression in the future.