feat: publicly export `UniqueIter` (#87)
This type is returned by a public method, but it isn't publicly
re-exported. This means that it cannot currently be named in downstream
code, and the documentation for its `Iterator` implementation is not
rendered in RustDoc. This is an oversight on my part, as the type was
intended to be public.

This commit fixes that, by publicly re-exporting `UniqueIter`. In
addition, I've added some improved documentation for `UniqueIter`, which
should help users to understand its guarantees.

Fixes #77