chore(ci): add job for everything passing (#89)
This way, we can use that as the required check for merging PRs, so we
don't have to depend on a bunch of stuff in the GitHub branch protection
settings. This means that the required checks are defined in the CI
config file, rather than in the UI, and if we change job names
or add new checks, we can add them to the requirements in `ci.yml`
instead of the repo settings.