chore(ci): don't require nightly to pass (#89)
This changes the nightly CI build to not be required. This is because
nightly Rust may contain regressions that incorrectly break our build.
If such a regression occurs, this is a Rustc bug rather than an issue
with a PR, and we should still be able to merge the PR. The purpose of
the nightly CI build is to warn us if something is broken on nightly so
that we can open issues against Rust upstream for nightly regressions,
rather than as a gatekeeper for merging PRs.