chore(docs): Fix docs links (#105)
Closes #69

As noted in #69, all hyperlinks to `slab` or `Slab` in the lib.rs
top-level doc comment link to https://crates.io/crates/loom. This seems
to be result of 2 errors dating back to the initial commit, relating to
this code:

```rs
//! - [`slab`]: Carl Lerche's `slab` crate provides a slab implementation with a
//!   similar API, implemented by storing all data in a single vector.
...
//! [`slab`]: https://crates.io/crates/loom
```

There seems to be 2 problems with this:

1. The markdown is set up to link `slab` to
   https://crates.io/crates/loom instead of
   https://crates.io/crates/slab. There is another part of the doc
   comment containing `//! [`loom`]: https://crates.io/crates/loom`, so
   I assume this was a copy-paste error.
2. The markdown is set up to link _all_ occurrences of `slab` to that,
   and `cargo doc` seems to do this in a case-insensitive way. This
   means that occurrences of `[`Slab`]` in the comment also get linked
   to that.

This PR fixes these 2 problems so that `Slab` links in the lib.rs docs
link to the `sharded_slab::Slab` item.